backend/requirements.txt
asgiref==3.7.2
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django>=3.2.0,<4.0.0
django-cors-headers>=3.0.0,<4.0.0
djangorestframework==3.14.0
google-api-core==2.14.0
google-auth==2.23.4
google-cloud==0.34.0
google-cloud-bigquery==3.13.0
google-cloud-core==2.3.3
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpcio==1.33.*
grpcio-status==1.59.3
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
proto-plus==1.22.3
protobuf>=3.0.0,<4.24.3   
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
rsa==4.9
six==1.16.0
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.0.7
